requirements.txt

pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
playwright==1.58.0
allure-pytest==2.15.3
python-dotenv==1.2.2
Faker